+++ src/components/cell-control/select.jsx
@@ -18,7 +18,6 @@ const Select = forwardRef(function Select(props, ref) {
     >
       <SelectControl
         classNames={{
-          container: styles.container,
           input: styles.input,
           invalidInput: styles.invalidInput,
         }}
@@ -31,7 +30,6 @@ const Select = forwardRef(function Select(props, ref) {
         ref={ref}
         required={props.required}
         validationMessage={props.validationMessage}
-        validationMessageTooltip
         value={props.value}
         wrapperRef={refs.container}
       >